+++ web/usuaris/src/__tests__/organisms/UsersTable.test.tsx
@@ -2,6 +2,7 @@ import { fireEvent, render, screen, waitFor } from '../test-utils'
 import { UsersTable } from '../../components/organisms'
 import { errorHandlers } from '../../__mocks__/handlers'
 import { server } from '../../__mocks__/server'
+import { UserStatus } from '../../types'
 
 afterEach(() => {
   server.resetHandlers()
@@ -62,6 +63,7 @@ describe('UsersTable', () => {
       <UsersTable
         filtersSelected={{
           itinerarySlug: 'frontend-react',
+          status: UserStatus.ACTIVE,
           startDate: '2023/11/06 00:00:00.000',
           endDate: '2023/11/06 00:00:00.000',
           name: 'marc',