+++ web/usuaris/src/components/molecules/index.ts
@@ -1,4 +1,5 @@
 export { DateRange, type TDateRange } from './DateRange'
 export { ItineraryDropdown } from './ItineraryDropdown'
 export { RolFilter } from './RolFilter'
+export { StatusDropdown } from './StatusDropdown'
 export { Table, type TTable } from './Table'