+++ web/usuaris/src/components/organisms/FiltersWidget.tsx
@@ -2,11 +2,12 @@ import { FC, useCallback } from 'react'
 import styled from 'styled-components'
 import { FlexBox, Search, dimensions } from '@itacademy/ui'
 import { useTranslation } from 'react-i18next'
-import { DateRange } from '../molecules'
+import { DateRange, StatusDropdown } from '../molecules'
 import { ItineraryDropdown } from '../molecules/ItineraryDropdown'
 import { RolFilter } from '../molecules/RolFilter'
 import { TRol } from '../../types/types'
-import { type TFilters, TItinerary } from '../../types'
+import { type TFilters, TItinerary, UserStatus } from '../../types'
+
 
 const FiltersContainer = styled(FlexBox)`
   width: 100%;
@@ -27,6 +28,10 @@ export const FiltersWidget: FC<TFiltersWidget> = ({ filters, setFilters }) => {
       setFilters({ ...filters, role: role?.slug })
   }
 
+  const handleStatus = (selectedStatus: UserStatus | undefined) => {
+    setFilters({ ...filters, status: selectedStatus })
+  }
+
   const handleSearch = useCallback(
     (searchValue: string) => {
       setFilters({
@@ -38,6 +43,7 @@ export const FiltersWidget: FC<TFiltersWidget> = ({ filters, setFilters }) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [filters]
   )
+
   const handleDates = (startDate: Date | null, endDate: Date | null) => {
     setFilters({
       ...filters,
@@ -54,6 +60,7 @@ export const FiltersWidget: FC<TFiltersWidget> = ({ filters, setFilters }) => {
     >
       <ItineraryDropdown handleItinerary={handleItinerary} />
       <RolFilter handleRole={handleRole} />
+      <StatusDropdown handleStatus={handleStatus} />
       <DateRange
         labelStartDate={t('Fecha de inicio')}
         labelEndDate={t('Fecha final')}