+++ web/usuaris/src/components/organisms/UsersTable/UsersTable.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { ColumnDef, createColumnHelper } from '@tanstack/react-table'
 import { Checkbox, Spinner, dimensions } from '@itacademy/ui'
 import { Table } from '../../molecules'
-import { TFilters, TUserData } from '../../../types'
+import { TFilters, TUserData, UserStatus } from '../../../types'
 import { icons } from '../../../assets/icons'
 import { useGetUsers, useUpdateUser } from '../../../hooks'
 import {
@@ -32,7 +32,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
 
   const { isLoading, isError, data: users } = useGetUsers(filters)
 
-  const [selectedStatus, setSelectedStatus] = useState<string | undefined>(
+  const [selectedStatus, setSelectedStatus] = useState<UserStatus | undefined>(
     undefined
   )
 
@@ -44,8 +44,9 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
     setFilters(filtersSelected)
   }, [filtersSelected])
 
-  const handleStatus = (id: string, status: string) => {
-    const updatedStatus = status === 'ACTIVE' ? 'BLOCKED' : 'ACTIVE'
+  const handleStatus = (id: string, status: UserStatus) => {
+    const updatedStatus =
+      status === UserStatus.ACTIVE ? UserStatus.BLOCKED : UserStatus.ACTIVE
 
     const updatedUser = {
       id,
@@ -72,7 +73,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
   const changeSelection = (
     e: ChangeEvent<HTMLInputElement>,
     id: string,
-    status: string
+    status: UserStatus
   ) => {
     if (e.target.checked) {
       setSelectedStatus(status)
@@ -97,13 +98,14 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
 
   const columHelper = createColumnHelper<TUserData>()
 
-  const columns: ColumnDef<TUserData, UserRole>[] = [
+  const columns: ColumnDef<TUserData, UserStatus, UserRole>[] = [
+
     columHelper.accessor('id', {
       header: '',
       cell: ({ row }) => {
         const id: string = row.getValue('id')
         const name: string = row.getValue('name')
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         let isDisabled: boolean | undefined
 
         if (selectedStatus && selectedStatus !== status) {
@@ -132,7 +134,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
       header: `${t('Nombre')}`,
       cell: ({ row }) => {
         const name: string = row.getValue('name')
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         let isDisabled: boolean | undefined
 
         if (selectedStatus && selectedStatus !== status) {
@@ -147,7 +149,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
       header: 'DNI/NIE',
       cell: ({ row }) => {
         const dni: string = row.getValue('dni')
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         let isDisabled: boolean | undefined
 
         if (selectedStatus && selectedStatus !== status) {
@@ -162,7 +164,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
       header: `${t('Especialización')}`,
       cell: ({ row }) => {
         const itineraryName: string = row.getValue('itineraryName')
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         let isDisabled: boolean | undefined
 
         if (selectedStatus && selectedStatus !== status) {
@@ -183,9 +185,8 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
         </CellStyled>
       ),
       cell: ({ row }) => {
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         let isDisabled: boolean | undefined
-        let displayedStatus: string = ''
 
         if (selectedStatus && selectedStatus !== status) {
           isDisabled = true
@@ -193,16 +194,9 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
           isDisabled = undefined
         }
 
-        if (status === 'PENDING') {
-          displayedStatus = t('pendiente')
-        } else if (status === 'BLOCKED') {
-          displayedStatus = t('bloqueado')
-        } else {
-          displayedStatus = t('activo')
-        }
         return (
           <DisabledStyled disabled={isDisabled}>
-            <StatusStyled status={status}>{displayedStatus}</StatusStyled>
+            <StatusStyled status={status}>{t(`${status}`)}</StatusStyled>
           </DisabledStyled>
         )
       },
@@ -211,7 +205,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
       header: `${t('Fecha alta')}`,
       cell: ({ row }) => {
         const createdAt = row.getValue('createdAt')
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         const formattedDate = new Date(createdAt as string).toLocaleDateString()
         let isDisabled: boolean | undefined
 
@@ -254,7 +248,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
         </ActionsHeader>
       ),
       cell: ({ row }) => {
-        const status: string = row.getValue('status')
+        const status: UserStatus = row.getValue('status')
         const id: string = row.getValue('id')
         let isDisabled: boolean | undefined
         let buttonTxt: string = ''
@@ -265,9 +259,9 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
           isDisabled = undefined
         }
 
-        if (status === 'PENDING') {
+        if (status === UserStatus.PENDING) {
           buttonTxt = t('Aceptar')
-        } else if (status === 'BLOCKED') {
+        } else if (status === UserStatus.BLOCKED) {
           buttonTxt = t('Desbloquear')
         } else {
           buttonTxt = t('Bloquear')