+++ web/usuaris/src/types/types.tsx
@@ -1,6 +1,12 @@
+export enum UserStatus {
+  ACTIVE = 'ACTIVE',
+  PENDING = 'PENDING',
+  BLOCKED = 'BLOCKED',
+}
+
 export type TFilters = {
   itinerarySlug?: string
-  status?: string
+  status?: UserStatus
   startDate?: string
   endDate?: string
   name?: string
@@ -27,7 +33,7 @@ export type TUpdatedUser = {
   name?: string
   password?: string
   role?: UserRole
-  status?: string
+  status?: UserStatus
   deletedAt?: string
   itineraryId?: string
 }
@@ -36,7 +42,7 @@ export type TUserData = {
   id: string
   name: string
   dni: string
-  status: string
+  status: UserStatus
   createdAt: string
   itineraryName: string
   role: UserRole