+++ web/usuaris/CHANGELOG.md
@@ -2,14 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.14.0] - 2024-05-02
+
+## [0.16.0] - 2024-05-09
 
 ### Added
 
 - Added Role column to UsersTable 
 - Added Roles filter feature
-- Add Test for Roles
   
+
+## [0.15.0] - 2024-05-08
+
+### Added
+
+- Added Status filter
+
+## [0.14.0] - 2024-05-06
+
+### Added
+
+- Refetch users after user status changes to update interface
+
 
 ## [0.13.0] - 2024-04-29